requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
geoalchemy2==0.12.5
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.3.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
packaging==21.3; python_version >= '3.6'
psycopg2-binary==2.9.4
pyparsing==3.0.9; python_full_version >= '3.6.8'
python-dotenv==0.21.0
sqlalchemy==1.4.42
